+++ src/utils/validation/index.ts
@@ -1,5 +1,3 @@
-export * from './stringIsInteger';
-export * from './stringIsFloat';
 export * from './maxDecimals';
 export * from './getIdentifierType';
 export * from './addressIsValid';